update docs and move gqrs to qrs module
