+++ docs/index.rst
@@ -4,9 +4,14 @@ wfdb
 Introduction
 ------------
 
-The native Python waveform-database (WFDB) package. A library of tools for reading, writing, and processing WFDB signals and annotations.
+The native Python waveform-database (WFDB) package. A library of tools
+for reading, writing, and processing WFDB signals and annotations.
 
-Core components of this package are based on the original WFDB specifications. This package does not contain the exact same functionality as the original WFDB package. It aims to implement as many of its core features as possible, with user-friendly APIs. Additional useful physiological signal-processing tools are added over time.
+Core components of this package are based on the original WFDB
+specifications. This package does not contain the exact same
+functionality as the original WFDB package. It aims to implement as many
+of its core features as possible, with user-friendly APIs. Additional
+useful physiological signal-processing tools are added over time.
 
 
 Development
