+++ docs/plot.rst
@@ -2,4 +2,4 @@ plot
 ----
 
 .. automodule:: wfdb.plot
-    :members: plot_record, plot_annotation, plot_all_records
+    :members: plot_items, plot_wfdb, plot_all_records
