+++ docs/wfdb.rst
@@ -1,10 +1,15 @@
 wfdb
-----
+====
 
-These core components are accessible by importing the `wfdb` package directly, as well as from their respective subpackages.
+These core components are accessible by importing the `wfdb` package
+directly, as well as from their respective subpackages.
+
+
+WFDB Records
+---------------
 
 .. automodule:: wfdb
-    :members: rdrecord, rdsamp, wrsamp, get_dbs, dl_database, dl_files, plot_record, plot_annotation, plot_all_records
+    :members: rdrecord, rdsamp, wrsamp
 
 .. autoclass:: wfdb.Record
     :members: wrsamp, adc, dac
@@ -12,5 +17,26 @@ These core components are accessible by importing the `wfdb` package directly, a
 .. autoclass:: wfdb.MultiRecord
     :members: multi_to_single
 
+
+WFDB Anotations
+---------------
+
+.. automodule:: wfdb
+    :members: rdann, wrann, show_ann_labels, show_ann_classes
+
 .. autoclass:: wfdb.Annotation
     :members: wrann
+
+
+Downloading
+-----------
+
+.. automodule:: wfdb
+    :members: get_dbs, get_record_list, dl_database
+
+
+Plotting
+--------
+
+.. automodule:: wfdb
+    :members: plot_items, plot_wfdb, plot_all_records
