+++ wfdb/__init__.py
@@ -1,11 +1,8 @@
-"""
-The native Python waveform-database (WFDB) package. A library of tools for reading, writing, and processing WFDB signals and annotations.
-"""
 from .io.record import (Record, MultiRecord, rdheader, rdrecord, rdsamp,
-    wrsamp, dl_database)
+                        wrsamp, dl_database)
 from .io.annotation import (Annotation, rdann, wrann, show_ann_labels,
-    show_ann_classes)
-from .io.download import get_dbs, dl_files
+                            show_ann_classes)
+from .io.download import get_dbs, get_record_list
 from .plot.plot import plot_items, plot_wfdb, plot_all_records
 
 from .version import __version__
