+++ wfdb/io/__init__.py
@@ -1,9 +1,6 @@
-"""
-The input/output subpackage contains classes used to represent WFDB objects, and functions to read, write, and download WFDB files.
-"""
 from .record import (Record, MultiRecord, rdheader, rdrecord, rdsamp, wrsamp,
-    dl_database, sig_classes)
+                     dl_database, sig_classes)
 from ._signal import est_res, wr_dat_file
 from .annotation import (Annotation, rdann, wrann, show_ann_labels,
-    show_ann_classes)
-from .download import get_dbs, dl_files
+                         show_ann_classes)
+from .download import get_dbs, get_record_list, dl_files
